+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Bug fixes:
 - Ensure that clicking on the input of an overweight (i.e. disabled) bAsset when minting
 does not enable the bAsset
 - Ensure that token balances and allowances are fetched as soon as they are available
+- Add `purpose` to `finalize` callback for transactions such that the callback has the 
+correct dependencies
 
 ## Version 1.1.3